+++ R/arc-select.R
@@ -60,6 +60,14 @@
 #'   fields = c("OBJECTID", "PlaceName"),
 #'   where = "TotalPopulation > 1000000"
 #' )
+#'
+#' # any Esri query parameter can be passed through `...`
+#' arc_select(
+#'   flayer,
+#'   fields = "StateAbbr",
+#'   geometry = FALSE,
+#'   returnDistinctValues = TRUE
+#' )
 #' }
 #' @returns An sf object, or a data.frame
 arc_select <- function(